rename report files for easier reference
The partners talk about them most often by their number
